Expand the fbin ingest example into a full optimized CAGRA-HNSW build
Rename OptimizedFbinIngestExample -> OptimizedCagraHnswBuildExample and
turn on every build-side optimization, so the example is a single golden
reference for building a large accelerated HNSW index rather than just an
ingestion pattern:

- Native flat buffering: size the codec's numInputVectors to each
  segment, streaming vectors into the native host matrix instead of a
  heap List<float[]> (single-segment; ~halves peak host memory).
- Automatic graph-build algorithm: HEURISTIC + AUTO_SELECT lets cuVS pick
  NN_DESCENT vs IVF_PQ by dataset size and auto-tune its parameters.
- Partitioned multi-segment build with a user-specified segment count:
    - sequential: K single-segment passes appended to one directory,
      peak host = one slice (N/K);
    - overlapped: a bounded pool (PIPELINE_DEPTH) builds segments into
      their own dirs with the GPU commit serialized on a permit, then
      hardlinks them into the final directory via addIndexes (no bulk
      copy). Peak host = depth * (N/K).

The class Javadoc states the assumptions behind the segment count: native
flat buffering forces one segment per slice, the GPU build is serialized
(more segments buy host-memory headroom + ingest overlap, not device
parallelism), and search fans out across all K segment graphs.

Adds a lucene-misc dependency (HardlinkCopyDirectoryWrapper) and updates
the examples README with the new name and 4-arg usage.
