+++ examples/README.md
@@ -40,14 +40,16 @@ To run the Index and Search on GPU example do:
 mvn clean install && java -Djava.util.logging.config.file=src/main/resources/logging.properties -cp target/examples-26.10.0-jar-with-merged-services.jar com.nvidia.cuvs.lucene.examples.IndexAndSearchonGPUExample
 ```
 
-To run the chunked `.fbin` ingestion example (reference pattern for efficiently streaming a large
-vector file into an accelerated HNSW index — open the file once, read sequential prefetched chunks
-that overlap the disk read with indexing, hold at most two chunks in memory, and reuse a single
-vector array across all documents) do:
+To run the optimized CAGRA-HNSW build example (reference pattern for efficiently building an
+accelerated HNSW index from a large `.fbin` with every ingest-side knob on — open the file once and
+stream sequential prefetched chunks that overlap the disk read with indexing, hold at most two chunks
+in memory, reuse a single vector array, size a native flat buffer per segment, auto-select the CAGRA
+graph-build algorithm, and optionally partition into K segments built sequentially or overlapped) do:
 
 ```sh
-mvn clean install && java -Djava.util.logging.config.file=src/main/resources/logging.properties -cp target/examples-26.10.0-jar-with-merged-services.jar com.nvidia.cuvs.lucene.examples.OptimizedFbinIngestExample
+mvn clean install && java -Djava.util.logging.config.file=src/main/resources/logging.properties -cp target/examples-26.10.0-jar-with-merged-services.jar com.nvidia.cuvs.lucene.examples.OptimizedCagraHnswBuildExample
 ```
 
-With no arguments it generates and indexes a small demo `.fbin`; pass a real file and chunk size as
-`... OptimizedFbinIngestExample <path-to.fbin> <chunkSizeMB>`.
+With no arguments it generates and indexes a small demo `.fbin` as a single segment; pass a real file,
+chunk size, segment count, and overlap flag as
+`... OptimizedCagraHnswBuildExample <path-to.fbin> <chunkSizeMB> <numSegments> <overlap:true|false>`.
